+++ ironelder/AndroidArchitecture/app/src/main/java/com/ironelder/androidarchitecture/component/BaseFragment.kt
@@ -6,14 +6,6 @@ import androidx.fragment.app.Fragment
 import com.ironelder.androidarchitecture.R
 
 abstract class BaseFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_main, container, false)
-    }
-
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         super.onCreateOptionsMenu(menu, inflater)
         menu.clear()
